# contracts/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Complete Inventory Management System API",
    "version": "1.5.0",
    "description": "نظام إدارة المخزون المتكامل - واجهة برمجة التطبيقات"
  },
  "servers": [
    {
      "url": "http://localhost:5002",
      "description": "Default server"
    }
  ],
  "paths": {
...
    "/api/auth/refresh": {
      "post": {
        "operationId": "post_auth_unified_refresh",
        "summary": "Refresh access token",
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "GenericSuccess": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
...
      "RefreshRequest": {
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}